% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_precip_stats.R
\name{daily.precip.stats}
\alias{daily.precip.stats}
\title{Return daily precipitation statistics for a site}
\usage{
daily.precip.stats(site)
}
\arguments{
\item{site}{Parameter of class character.
The NEON site data should be downloaded for.}
}
\value{
A list of min, mean and max precipitaiton
values at the site, in milimeters
}
\description{
This function calculates the daily miniumum, mean, and maximum
precipitation values for a site over its period of record.
}
\examples{
\dontrun{
cper=Z10::daily.precip.stats(site = "CPER")
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rhlee@colorado.edu}\cr
}
