% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_soil_temp_stats.R
\name{daily.soil.temp.stats}
\alias{daily.soil.temp.stats}
\title{Return daily soil temperature statistics}
\usage{
daily.soil.temp.stats(site)
}
\arguments{
\item{site}{Parameter of class character.
The NEON site data should be downloaded for.}
}
\value{
A list of min, mean and max net solar radiation
values at the site, in watts per meter squared
}
\description{
This function calculates the daily miniumum, mean, and maximum
temperature values for a site over its period of record for soil sensors located
in plot 1 of the site, in the lowest available A horizon and the
uppermost B horizon (excluding AB horizons)
}
\examples{
\dontrun{
cper=Z10::daily.soil.temp.stats(site = "CPER")
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rhlee@colorado.edu}\cr
}
