% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAT.R
\name{mat}
\alias{mat}
\title{Return the Mean Annual Temperature at a site}
\usage{
mat(site)
}
\arguments{
\item{site}{Parameter of class character.
The NEON site data should be downloaded for.}
}
\value{
A list of min, mean and max temperature
values at the site, in centigrade
}
\description{
This function calculates the miniumum, mean, and maximum
temperature values for a site over its period of record.
}
\examples{
\dontrun{
cper=Z10::mat(site = "CPER")
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rhlee@colorado.edu}\cr
}
