% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_litter_isotopes.r
\name{site.litter.isotopes}
\alias{site.litter.isotopes}
\title{Return Mean Delta Values of Stable Isotopes in Litterfall}
\usage{
site.litter.isotopes(site)
}
\arguments{
\item{site}{Parameter of class character.
The NEON site data should be downloaded for.}
}
\value{
A list of min, mean and max net solar radiation
values at the site, in watts per meter squared
}
\description{
This function calculates the mean delta values for
nitrogen-15 and carbon-13 isotopes over the period of record at a site.
}
\examples{
\dontrun{
cper=Z10::site.litter.isotopes(site = "SCBI")
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rhlee@colorado.edu}\cr
}
