% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_rad_stats.R
\name{daily.rad.stats}
\alias{daily.rad.stats}
\title{Return daily solar radiation statistics for a site}
\usage{
daily.rad.stats(site)
}
\arguments{
\item{site}{Parameter of class character.
The NEON site data should be downloaded for.}
}
\value{
A list of min, mean and max net solar radiation
values at the site, in milimeters
}
\description{
This function calculates the daily miniumum, mean, and maximum
net solar radiation values for a site over its period of record.
}
\examples{
\dontrun{
cper=Z10::daily.rad.stats(site = "CPER")
}
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rhlee@colorado.edu}\cr
}
